requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
shap>=0.41.0
streamlit>=1.22.0
discord.py>=2.0.0
PyMuPDF>=1.20.0
beautifulsoup4>=4.11.0
requests>=2.28.0
transformers>=4.25.0
pyyaml>=6.0.0
matplotlib>=3.5.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
isort>=5.10.0
pylint>=2.15.0
python-dotenv>=0.20.0
joblib>=1.2.0
backoff>=2.2.0